Move towards type checker
