+++ golem-rib/src/inferred_type/unification_result.rs
@@ -1,0 +1,12 @@
+use crate::InferredType;
+
+pub enum UnificationResult {
+    Success(InferredType),
+    Failed(String)
+}
+
+impl UnificationResult {
+    pub fn unified(inferred_type: InferredType) -> UnificationResult {
+        UnificationResult::Success(inferred_type)
+    }
+}
